requirements.txt
amqp==5.3.1
asgiref==3.11.0
beautifulsoup4==4.14.3
billiard==4.2.4
boto3==1.42.25
botocore==1.42.25
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
cssselect2==0.8.0
Django==6.0
gunicorn==23.0.0
django-allauth==65.13.1
django-htmx==1.27.0
django-multitenant==4.1.1
django-storages==1.14.6
django_celery_results==2.6.0
fonttools==4.61.1
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.6.2
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
pyOpenSSL==25.3.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qrcode==8.2
redis==7.1.0
requests==2.32.5
rules==3.5
s3transfer==0.16.0
satcfdi==4.9.1
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
tabulate==0.9.0
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
xlsxwriter==3.2.9
zopfli==0.4.0
